Fix bud issue with 2 line Dockerfile
We were checking for single line Dockerfile to ensure
that they are actually built, but the condition was
also applying to two line Dockerfiles causing those
ones to not build properly. Fixed the condition and
now single line as well as 2 or more line dockerfile
build as expected.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>

Closes: #1315
Approved by: rhatdan