Fix buildah commit compress by default
The change to switch over to cobra had the effect of always setting
compress to imagebuildah.Uncompressed for the commit and bud commands.
This is because the conditional to set it to Gzip always returns false
regardless of whether --disable-compression is set or not.

This change makes the compression flag logic in cmd/buildah/bud.go and
cmd/buildah/commit.go consistent with the logic in
cmd/buildah/push.go[1] (which shows the correct behaviour)

Signed-off-by: Steve Baker <sbaker@redhat.com>