healthcheck start-period: update documentation
A re-reading of the spec doesn't agree with our docs - the healthcheck
start period isn't a delay in the first invocation of the healthcheck
command, but is instead a grace period where failures aren't counted as
indications of problems.  Instead, during that grace period, failures
are attributed to the container not yet having finished starting.  Once
the healthcheck command succeeds, the engine should start counting
failures.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>