imagebuildah: split Executor and per-stage execution
When building images using multiple stages, we were using a shallow copy
of the Executor, depending on some fields to be shared between all
instances, while replacing others fields with per-stage values.

This patch makes that easier for me to reason about by introducing a
StageExecutor which holds the stage-specific information, and which can
hold a reference to the main Executor, which now coordinates stages
using StageExecutors rather than Executors.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>