imagebuildah: drop the duplicate containerIDs list
StageExecutors keep their own lists of the IDs of containers that they
create, so we don't need to merge them into a single list in the
Executor that we'll consult during clean up.  As long as we make sure we
only ever try to delete a given stage's containers once, we can just use
their lists directly, and avoid duplicating the information.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>