vendor latest openshift/imagebuilder
openshift/imagebuilder had a dependency on a deprecated package that is
now forked into imagebuilder to allow using more recent versions of
docker/docker.

Also vendor the latest docker/docker to a) demonstrate that imagebuilder
does not force using an old version docker/docker and to b) pull in
fixes and updates to our dependencies on docker/docker.

See https://github.com/openshift/imagebuilder/pull/125 for more details
on the imagebuilder change.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>