Print commitID with --layers
Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>

When the --layers option is used, the final imageID was
not being printed in the final line.  It was printed two
lines prior.  We still have to print the one two lines up
(at least without major modifications), but now we print
it in the last line as it does for all other circumstances
so it can be grabbed by scripts.

Fixes: #1237

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>