imagebuildah.Build: consolidate cleanup logic
Consolidate all of the logic related to cleaning up after a build into a
helper function in Build().  This lets us ensure that we always clean up
images after cleaning up containers, so that we don't have to worry
about ordering when some images may be in use by containers because we
wanted access to a root filesystem for a stage whose image was reused.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>