imagebuildah: reorganize stage and per-stage logic
Move the preparation for the first container and the commit of the final
image in a single-stage build into the per-stage Execute() function.

For multi-stage builds, only create a layer for ADD, COPY, RUN, and the
final instruction in a stage, which changes the step counts that we
output.

Stop removing the final image for intermediate stages, since they're
expected to be present as cached images, even if they don't end up as
ancestors of the final image, except for single-stage builds, which we
document as not populating the cache.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #1473
Approved by: rhatdan