imagebuildah: defer pulling images for COPY --from
When handling COPY --from instructions, defer pulling images and
instantiating containers so that we can read content from their root
filesystems until we get to the instruction.

Have the main Executor keep track of the Builders that we create for
that purpose, and have it clean them up at the end of the build, as it
did when they were treated as their own stages.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>