Travis: do not use traviswait
The traviswait script has plenty of undesired side-effects, mainly but
not limited to not printing std{err,out} when hitting the time out.
This is especially painful when the tests hit a deadlock, rendering it
useless for debugging the deadlock.  Besides that, it also slows down
the tests causing frequent time outs.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>

Closes: #1488
Approved by: rhatdan