Travis CI: fail fast, upon error in any step
Use 'set -e' to exit immediately upon any failure: there is,
e.g., no reason to run BATS tests if the compile fails.

Background: in my brief experience here, the Go:tip jobs
inevitably fail. Although these are marked as allowed to
fail, I tried looking at the results anyway to see if the
failure was my fault; it took me a while to figure out
that the go build was failing, yet the CI job was plodding
on regardless. This seems wrong on infinite levels, and
a number of other people agree:

   https://github.com/travis-ci/travis-ci/issues/1066

In addition, remove the entire 'after_script' section.
It contained a step that (a) doesn't actually work,
and (b) probably belongs in a separate CI job on its own.

Signed-off-by: Ed Santiago <santiago@redhat.com>

Closes: #1494
Approved by: rhatdan