Don't fail bud --layers on dangling symlink
If there is a dangling symlink when building with --layers,
don't fail and continue the build process.
It would fail because it would try to stat the file to check
whether it has been modified since the previous build. And since
it would be a dangling symlink, the stat would fail hence causing
the build to fail.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>