Move buildah options into buildah/pkg/options for easier podman windows compile
Currently the podman-remote command needs access to the build options and the commit
options in the client.  Since these options are defined in buildah main line code, go
ends up sucking in the code and it makes it very difficult to compile this out on
Windows and Mac Platforms.

Bu moving the code to buildah/util subdir, we should be able to avoid these problems.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>