allow portions of buildah to build for windows
the podman remote client for windows needs to be able to compile certain
portions of buildah for windows.  this seems like the safest way to not
break api for buildah.

as podman matures, we might be able to avoid this by not compiling some
of the core pieces of libpod.  that is a longer term process however.

Signed-off-by: baude <bbaude@redhat.com>