imagebuildah: addHelper: handle symbolic links
When filtering copying/adding operations to comply with a .dockerignore
file, we should directly handle copying directories and symbolic links.
This update still returns an error when attempting to copy
non-directory, non-regular items.

Make DockerIgnoreHelper() private, since it's not being used by packages
that use us, and I expect it to be removed when this gets refactored.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #1583
Approved by: rhatdan