config: expand variables only at the command line
Skip our own attempt to expand variables in the SetEnv API, since we
need to let the Dockerfile parser be the only place that happens when
we're building using a Dockerfile.

Move the expansion logic to the CLI, and have it log an error when it
encounters a variable that it can't expand.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #1607
Approved by: TomSweeneyRedHat