Add ability to skip secrets in mounts file
This commit changes the behavior for mounting secrets from a file like
`/etc/containers/mounts.conf`.

Before this commit, lines which do not contain a possible mount (like
comments) where included in the the returned slice of `getMounts()`.
This means that the returned result is not correct and higher level
functions like `addSecretsFromMountsFile()` fail because of that
comment.

Now, invalid lines are simply skipped, which allows the end-user to keep
their additional information within the mounts files.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>