Added '-' syntax to remove volume config option
When using the --volume flag, ojects ending in "-" are checked if they exist on disk. If they exist on disk they are added as a volume. Otherwise, if the volume is already config'd, that volume is removed

examples:
if /testvol exists:
"--volume /testvol-" removes /testvol from the config.

if /testvol- exists:
"--volume /testvol-" adds /testvol- from the config.

Signed-off-by: Ashley Cui <ashleycui16@gmail.com>