add: fix slow copy with no excludes
when there are excludes defined (such as from .dockerignore), we take
a slow path and walk each file in the directory.  If the files doesn't
match any exclusion pattern then it is copied into the container.
This is slow as each file requires buildah to re-exec and copy it from
a chroot environment.

When there are no excludes defined we can take a faster path and copy
the entire directory as a single re-exec operation.

Closes: https://github.com/containers/buildah/issues/1714

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #1715
Approved by: rhatdan