unshare: add a --mount flag
Add a --mount flag to `buildah unshare`, so that the command it runs can
have a container's root filesystem mounted, and the location where it's
mounted provided in its environment.

This is primarily aimed at cases where `buildah unshare` is used to run
one-liner commands, as proper scripts can simply call `buildah mount`
themselves.

Users still have to be careful about quoting references to the environment
if the command needs to be executed by a shell (for example, if it's run
as `buildah unshare --mount foo sh -c 'ls $foo'`), so that the shell
that's invoking `buildah unshare` doesn't try to expand the variable's
value first.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>