Add --digestfile and Re-add push statement as debug
We removed the "Succesfully pushed ..." statement from the end of
the `buildah push` command in #1744.  However we've had a request
for that information still so I'm re-adding it as a debug statement
and we can discuss if there's a better way or not.

Also add the --digestfile option.  With it the user will be able
to specify a file to write the digest of the pushed image to for
reference.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>

Closes: #1799
Approved by: rhatdan