Fix directory pull image names
This is a breaking change and modifies the resulting image name when
pull from an directory via `oci:...` or `dir:...`.

Without this patch, the image names pulled via a local directory got
prefixed with `docker.io/{library/}`, which is not correct.

We now use either the full path to the image, or the relative path as
image name, but prefixed with `localhost` to indicate that the image is
not being pulled from a remote location.

Fixes: https://github.com/containers/buildah/issues/1797

Signed-off-by: Sascha Grunert <sgrunert@suse.com>