Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

This renames BUILD_INFO to the standard name
so that distributions do not have to discover the BUILD_INFO variable,
similar to https://github.com/cri-o/cri-o/pull/2643

Signed-off-by: Bernhard M. Wiedemann <bwiedemann@suse.de>