manifests: add the module
Wrap the pkg/manifests API in a version which knows how to load and save
manifest lists and information about where the various instances are
found in local storage.

Add an ImageSource implementation that can use a manifest list and
instance information to create a grab bag of all of the blobs known to
all of the instances.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>