Fix --build-args handling
When passing an environment variable to a Containerfile using an `ARG` command,
the dry-run processing  for `COPY` and `ADD` isn't resolving the environment
variable for the source and is causing a failure.

Now convert the source if it's an environment variable to its value and then
proceed on.  The destination doesn't have a similar issue at this point
but we'll convert it too just in case.

Addresses: #1871

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>

Closes: #1874
Approved by: nalind