Ensure bud remote context cleans up on error
During the unpacking of remote context for '-'(stdin), child commit,
the possibility for the remote context processing to leave behind a
stale context directory in the temporary directory `/var/run` was called
out.

As the context is unpacked into a temporary directory, it should be safe
to remove the directory and all files under it.  The files under the
path should only ever be the content of an unpacked tar, a checked out
git repository, or a lone "Dockerfile".

Signed-off-by: Sachi King <nakato@nakato.io>