Add squash-all for podman build
Add a `--squash-all` flag that will be unique to `podman build` and will not
be used by Buildah.  `buildah bud --layers=false` is equivalent to `docker build --squash`.  When a podman build command includes the `--squash` option, podman will tranlate that to `--layers=false` for the build build api code.  A `docker build` call is equivalent to `podman build` or `podman build --layers=true` and that's not changing.

When podman build uses the `--squash-all` command, it will turn on the --squash option for the buildah bud api to have only one layer returned by the buildah bud code.  There's not a Docker equivalent for this at the moment.

I'll hopefully be spinning up a Podman PR in the next day or two that will use this new option.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>