Add clarification to the Tutorial for new users
This is a partial fix for #1895.  This focuses on (hopefully) helping a
user get into Buildah faster when they first hit the buildah.io site and
click down through the tutorials button.  It also clarifies what's currently
out of scope for Buildah, adds some titles to make links easier and a bit more.

I expect I'll either be adding to this or creating a follow on PR to address more
of the notes in #1895.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>