Fix another broken test: copy-url-mtime
The newer-than and older-than test was failing (desired result)
but it's because the command had a syntax error (not desired)!

Rewrite that section a little, making it more robust (actually
confirming that mtime is preserved, versus simple mtime match
because of time of execution) and also with more readable
diagnostic in case of error.

Signed-off-by: Ed Santiago <santiago@redhat.com>