Handle missing equal sign in --from and --chown flags for COPY/ADD
When buildah bud runs without layers (default), it hands off the parsing
of the Dockerfile steps to imagebuilder.  However if the the --layers
option is in use, then we start parsing the steps before we hand stuff
off to imagebuilder.

The COPY --chown, ADD --chown, and COPY --from commands in a Dockerfile
were failing if there was not an '=' sign after them like "COPY --chown=2374:3256".
This moves the parsing into the code that handles the layer processing.

Fixes: #1984

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>

Closes: #1997
Approved by: vrothberg