Add the --from option to bud
When processing a Dockerfile supplied to the bud command,
replace any image name after the FROM statement with the
value passed in to the --new-from option.  This will not
change the value in the on-disk version of the Dockerfile,
only the one in memory for the lifetime of the build
process.

If the image name is a variable, do not replace that.

I also found a number of missing params in the completions
and a small alpha ordering issue in the man page.

Addresses: #2007

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>