chroot: handle slightly broken seccomp defaults
When a seccomp rule includes multiple equality checks for the same
argument for a syscall, they can never ALL be satisfied.  Because that's
how they're supposed to be treated, libseccomp returns an error when we
try to add them as part of the same conditional rule.  Try to detect
this exact case, and if we detect it, treat each condition as its own
rule.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>