Adjust copy destination for linked tar files in ADD
When we are given a linked tar file as a source file, we were setting the
destination to the original destination plus the name of the link.
This created a directory using the links name and then everything was
copied under that directory.  The link name should not be used as part
of the directory structure.

Fixes: #2102

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>

Closes: #2112
Approved by: rhatdan