buildahimage: specify fuse-overlayfs mount options
tweak the storage.conf file to use fuse-overlayfs specific mount
options.  In particular, disable any sync operation on the files on
fuse-overlayfs.

sync is an expensive operation and there is no gain in sync'ing files
when doing a build as if the build fails we don't store the
intermediate result.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>