selinux spc test: fix CI breakage
The SELinux spc test is breaking under some CI circumstances in
Fedora 30; most notably, a CI-built buildah (with type != bin_t)
is failing with:

       FAIL: container context matches our own
   expected: 'system_u:system_r:unconfined_service_t:s0'
     actual: 'system_u:system_r:spc_t:s0'

It is not clear to me what the initial purpose of this test
is, since (I believe) the container context should always
be spc_t regardless of the calling process context; likewise,
role should always be system_r.

This commit tweaks the test to check role/type/range
individually.

Needed for #2115

Signed-off-by: Ed Santiago <santiago@redhat.com>