manifest add: always read the individual image manifest
Always read the manifest of the individual image that we're adding to a
list, so that we also read its MIME type.  When we weren't doing this
before, when we added an individual image from another list, we'd record
the type of that other list as the image's type.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #2140
Approved by: TomSweeneyRedHat