Show error on copied file above context directory in build
When a COPY command in a Container file looked like "../file.txt"
SecureJoin would secure the file by lopping off the "../".
However, the code would then append that file name to the passed in
context directory and look for the file.  That would fail as in most
cases there was no `{context-dir}/file.txt`, rather the file was at
`{context-dir}/../file.txt`.  Using a relative directory like this
outside of the context directory can be a security risk.  Docker
doesn't allow it nor should we.

This change now errors out when a file that starts with `../` is
presented as a copy from target.

Addresses: https://github.com/containers/libpod/issues/5035

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>